+++ Data/Data.lua
@@ -70,8 +70,6 @@ end
 
 local LoadBlizzardTabAchievements;
 local function PostLoadOnPlayerLogin(self, start)
-    -- self.ExportedAchievements.Load(self.AchievementIds);
-
     local custom = LibStub("AceConfigRegistry-3.0"):GetOptionsTable(addon.Metadata.Prefix .. "_Layout", "cmd", "KROWIAF-0.0").args.Summary.args.Summary.args.NumAchievements; -- cmd and KROWIAF-0.0 are just to make the function work
     custom.max = #self.AchievementIds;
 
@@ -104,7 +102,6 @@ function data:LoadOnPlayerLogin()
     KrowiAF.CreateBuildVersions();
 
     self:RegisterAchievementDataTasks();
-    self.ExportedAchievements.RegisterTasks(self.Achievements);
     self.ExportedCategories.RegisterTasks(self.Categories, self.Achievements, addon.Tabs);
     self:RegisterEventDataTasks();
     self.ExportedUiMaps.RegisterTasks(self.Maps, self.Achievements);